refactor: 단일 책임 원칙 적용하여 CommandLineInterface를 ThreadHandler 클래스로 나눔
